feat:  측정 결과 조회 기능 구현

See merge request s12-final/S12P31S209!133

+++ Android/app/src/main/java/com/example/myapplication/data/dto/response/measure/Temperature.kt
@@ -2,5 +2,5 @@ package com.example.myapplication.data.dto.response.measure
 
 data class Temperature(
     val level: String,
-    val value: Double
+    val value: Float
 )
